Update: update shoppingCart layout to fix bug
